requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.9
django-adminlte-ui==1.5.1
django-ckeditor==6.4.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-treebeard==4.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.33.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
MarkupPy==1.14
mysql-client==0.0.1
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.1.2
pyphen==0.12.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tinycss2==1.1.1
urllib3==1.26.9
weasyprint==54.3
webencodings==0.5.1
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
